+++ packages/af-utils-react-virtual-headless/src/models/ResizeObserver/index.js
@@ -1,9 +1,0 @@
-class FakeResizeObserver {
-    observe() {}
-    unobserve() {}
-    disconnect() {}
-}
-
-export default typeof ResizeObserver === "undefined"
-    ? FakeResizeObserver
-    : ResizeObserver;