+++ packages/af-utils-react-virtual-headless/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/react-virtual-headless
 
+## 0.0.12
+
+### Patch Changes
+
+-   separate bundles for client/server
+
 ## 0.0.11
 
 ### Patch Changes